Add MicroProfile Context Propagation subsystem and tests
This includes a SmallRye snapshot to get
https://github.com/smallrye/smallrye-context-propagation/pull/113 which
addresses a problem in the TCK when running in a multi-classloader
environment.

Adds support for the Application ThreadContext type, which is not
provided by SmallRye.

TODO: Support for the Security ThreadContext type. What Elytron expects
does not presently look like it maps cleanly to the reactive world.