Add TCKs
Add import of META-INF/ for archives which may include beans.xml. This
is needed following https://issues.redhat.com/browse/WFLY-13641 /
  https://github.com/wildfly/wildfly/pull/13406

Reactive Messaging TCK fails when everything is provisioned but works
when running a slimmer server. We need a new SmallRye version with the
hook to turn off context propagation to get it working.