[WFLY-14798] Upgrade reactive messaging to 3.6.0 and add Kafka API module
The Kafka API is tracked in WFLY-14932, and will be tested as part of that.
It is however a runtime dependency of SmallRye Reactive Messaging 3.6.0 onwards.

Messages get sent on different Kafka partitions, so the assumption that there is an
absolute ordering of messages is too naive. Rather they are ordered within each
partition. We will address that as part of the Kafka API implementation in WFLY-14932